+++ src/reducers/notification/reducer.ts
@@ -5,8 +5,8 @@ import { NotificationState } from "./state";
 
 export type NotificationActions = ActionType<typeof actions>;
 const initState: NotificationState = {
-  pending: { items: [] },
-  transaction: { items: [] },
+  pending: [],
+  transaction: [],
 };
 
 export function notificationReducer(
@@ -14,6 +14,16 @@ export function notificationReducer(
   action: NotificationActions,
 ): NotificationState {
   switch (action.type) {
+    case "ADD_TRANSACTION":
+      return {
+        ...state,
+        transaction: [...state.transaction, action.payload],
+      };
+    case "ADD_PENDING_TRANSACTION":
+      return {
+        ...state,
+        pending: [...state.pending, action.payload],
+      };
     default:
       return state;
   }
